test(cli): seal queued paste fixture before completion delivery
Problem: the coverage job intermittently failed while removing the queued-paste test's temporary session directory.
Root cause: TurnDone reached the test before the controller's completion tail attempted another durable-inbox dispatch, allowing disk access to overlap fixture cleanup.
Fix: close the single-turn controller inside the TurnDone sink before publishing completion. Keep the real persistent inbox and every queue/input assertion, remove the redundant runner polling, and extract the case from the oversized TUI test file.
Verification: the regression passes 100 repetitions with race detection and coverage; the complete CLI package coverage test and repolint pass. No production lifecycle or persistence semantics change.
